Add fish support
I've reused the prompt for bash/zsh for simplicity's sake. This could
definitely be improved in the future by having a more generic structure
for constructing the prompt based on the output of the `kubie info`
commands and a set of colors.

This change was verified with my fish prompt based on starship.rs, and
while it works, it's not prepending it on the same line. This is because
of the starship prompt is outputting a clearing ANSI escape code. (It
would be possible to check for the existence of that code and inject the
prompt after it, effectively ending up where one would expect the prompt
to be, though I'm not convinced that type of logic belongs in Kubie.)
The outcome is exactly the same as in
[this
comment](https://github.com/sbstp/kubie/issues/18#issuecomment-609477575).

The right-side prompt works as expected, though. I reused the zsh RPS1
setting for determining whether right-side prompt is enabled. It would
feel redundant to have one right-prompt setting per shell, but OTOH the
current name is hard-tied to zsh.