Run fish as an interactive login shell
Since kubie is used in an interactive context, the proper way to spawn a
shell should be to pass in the interactive and login flags to it (I
think). I noticed that ctrl+c for instance behaved weirdly in fish when
not setting the spawned shell as interactive and login.