Add support for unsetting namespace
`kubie ns -u` or `kubie ns --unset` will unset the currently active
namespace. It also does not set namespace at all by default (matching
e.g. `kubectx`) behavior. The original motivation was to get rid of my
starship prompt saying `context:default`; with this change I can now
have the prompt display just `context` and that will be the default for
new contexts without session history. Additionally, it allows a greater
set of legal kubeconfigs (i.e. with a null current namespace).

The built-in prompt is left as-is with `[context|default]` when the
namespace is unset. This matches the `kubens --current` behavior (and
doesn't change the Kubie prompt's current behavior).