document export subcomannd
The `export` subcommand was added in #125, but was not documented with
clap or in the project readme. As a result, it's not clear kubie can be
used alongside other tools such as direnv.

This changeset adds the docstrings for clap, rephrases the documentation
for the arguments, and includes it in the list of subcommands in the
project readme.